dax_momentum/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dax-momentum"
version = "0.1.0"
description = "Momentum-basierte Vorhersagepipeline für DAX-Aktien"
authors = [{name = "DAX Momentum Team"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.1",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "torch>=2.0",
    "refinitiv-data>=1.0; platform_system != 'Windows'",
    "pyyaml>=6.0",
    "loguru>=0.7",
    "python-dotenv>=1.0",
    "tqdm>=4.0",
    "apscheduler>=3.9",
    "pyarrow>=14.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
known_third_party = ["loguru", "numpy", "pandas", "refinitiv", "sklearn", "torch", "yaml"]
